[model] Change: Move CGRect, FlipMode, RotationMode into package com.guhungry.photomanipulator.model
