+++ photomanipulator/src/main/java/com/guhungry/photomanipulator/model/RotationMode.kt
@@ -1,4 +1,4 @@
-package com.guhungry.photomanipulator
+package com.guhungry.photomanipulator.model
 
 enum class RotationMode(val degrees: Float) {
     None(0f),
